# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the UI for the DomainScraper endpoints.  You can use this to test the calls before writing scripts.",
    "version": "1.0.0",
    "title": "DomainScraper"
  },
  "schemes": ["http"],
  "paths": {
    "/domain_data": {
      "get": {
        "tags": ["Domain Crawler"],
        "summary": "Detects technologies and social media used on a domain",
        "description": "Detects the technology and social media on a domain in real time.",
        "operationId": "DomainData",
        "produces": ["application/json"],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "Domain without protocol or path",
          "required": true,
          "type": "string",
          "format": "string"
        }],
        "responses": {
          "200": {
            "description": "success",
            "examples": {
              "response": {
                "domain": "test.com/",
                "status_code": 200,
                "data": {
                  "accounting": [],
                  "advertising_newtworks": [],
                  "analytics": [],
                  "blogs": []
                }
              }
            }
          },
          "500": {
            "description": "error"
          }
        }
      }
    },
    "/page_rank": {
      "get": {
        "tags": ["Page Rank"],
        "summary": "Gets the open page rank for up to 100 domains at a time.",
        "description": "Gets the open page rank for up to 100 domains at a time.",
        "operationId": "PageRank",
        "produces": ["application/json"],
        "parameters": [{
          "name": "urls",
          "in": "query",
          "description": "comma separated",
          "required": true,
          "type": "string",
          "format": "string"
        }],
        "responses": {
          "200": {
            "description": "success",
            "examples": {
              "response": {
                "domain": "test.com/",
                "status_code": 200,
                "data": {
                  "test.com": {
                    "page_rank": 100
                  }
                }
              }
            }
          },
          "500": {
            "description": "error"
          }
        }
      }
    },
    "/dns_info/ns": {
      "get": {
        "tags": ["DNS Info"],
        "summary": "Gets DNS information for a domain.",
        "description": "Gets DNS information for a domain.",
        "operationId": "DNSNSInfo",
        "produces": ["application/json"],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "Domain without protocol or path",
          "required": true,
          "type": "string",
          "format": "string"
        }],
        "responses": {
          "200": {
            "description": "success",
            "examples": {
              "response": {
                "domain": "test.com/",
                "status_code": 200,
                "data": {
                  "test.com": {
                    "page_rank": 100
                  }
                }
              }
            }
          },
          "500": {
            "description": "error"
          }
        }
      }
    },
    "/dns_info/cname": {
      "get": {
        "tags": ["DNS Info"],
        "summary": "Gets DNS information for a domain.",
        "description": "Gets DNS information for a domain.",
        "operationId": "DNSCNAMEInfo",
        "produces": ["application/json"],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "Domain without protocol or path",
          "required": true,
          "type": "string",
          "format": "string"
        }],
        "responses": {
          "200": {
            "description": "success",
            "examples": {
              "response": {
                "domain": "test.com/",
                "status_code": 200,
                "data": {
                  "test.com": {
                    "page_rank": 100
                  }
                }
              }
            }
          },
          "500": {
            "description": "error"
          }
        }
      }
    },
    "/whois_info/registrar": {
      "get": {
        "tags": ["WhoIs Info"],
        "summary": "Gets whois information for a domain.",
        "description": "Gets whois information for a domain.",
        "operationId": "whois_registrar_info",
        "produces": ["application/json"],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "Domain without protocol or path",
          "required": true,
          "type": "string",
          "format": "string"
        }],
        "responses": {
          "200": {
            "description": "success",
            "examples": {
              "response": {
                "domain": "test.com/",
                "status_code": 200,
                "data": {
                  "test.com": {
                    "page_rank": 100
                  }
                }
              }
            }
          },
          "500": {
            "description": "error"
          }
        }
      }
    },
    "/whois_info/company_name": {
      "get": {
        "tags": ["WhoIs Info"],
        "summary": "Gets whois information for a domain.",
        "description": "Gets whois information for a domain.",
        "operationId": "whois_company_info",
        "produces": ["application/json"],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "Domain without protocol or path",
          "required": true,
          "type": "string",
          "format": "string"
        }],
        "responses": {
          "200": {
            "description": "success",
            "examples": {
              "response": {
                "domain": "test.com/",
                "status_code": 200,
                "data": {
                  "test.com": {
                    "page_rank": 100
                  }
                }
              }
            }
          },
          "500": {
            "description": "error"
          }
        }
      }
    }
  }
}